% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble.R
\name{as_tibble.BasicRaster}
\alias{as_tibble.BasicRaster}
\title{Convert a Raster to a tibble.}
\usage{
\method{as_tibble}{BasicRaster}(x, cell = TRUE, dim = TRUE,
  split_date = TRUE, ...)
}
\arguments{
\item{x}{a RasterLayer, RasterStack or RasterBrick}
}
\value{
a tibble
}
\description{
Convert a Raster to a tibble.
}
\examples{
as_tibble(raster::raster(volcano))

as_tibble(setZ(raster::raster(volcano), Sys.Date()), cell = TRUE)
}
