% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble.R
\name{as_tibble}
\alias{as_tibble}
\alias{as_tibble.BasicRaster}
\title{Convert a Raster to a tibble.}
\usage{
\method{as_tibble}{BasicRaster}(x, cell = TRUE, dim = TRUE,
  split_date = TRUE, ...)
}
\arguments{
\item{x}{a RasterLayer, RasterStack or RasterBrick}

\item{cell}{logical to include explicit cell number}

\item{dim}{logical to include slice index}

\item{split_date}{logical to split date into components}

\item{...}{unused}
}
\value{
a tibble
}
\description{
Convert a Raster to a tibble.
}
\examples{
library(tibble)
as_tibble(raster::raster(volcano))
library(raster)
as_tibble(setZ(raster::raster(volcano), Sys.Date()), cell = TRUE)
}
