% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary.r
\name{boundary}
\alias{boundary}
\title{Create a boundary polygon by tracking around coordinates stored in a RasterStack}
\usage{
boundary(cds)
}
\arguments{
\item{cds}{two-layer Raster}
}
\description{
The first layer is treated as the X coordinate, second as Y.
}

