% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer_extent.r
\name{bufext}
\alias{bufext}
\title{Deprected tabularaster function}
\usage{
bufext(e1, e2)
}
\arguments{
\item{e1}{input \code{\link[raster]{extent}}}

\item{e2}{grain size}
}
\description{
Ensure a raster extent aligns to whole parts.
}
\details{
This function is deprecated, please use \code{spex::buffer_extent}.
}
\examples{
library(spex)
library(raster)
buffer_extent(extent(0.1, 2.2, 0, 3), 2)
}
