% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellnumbers.R
\name{cellnumbers}
\alias{cellnumbers}
\alias{pcellnumbers}
\alias{pwcellnumbers}
\title{Title}
\usage{
cellnumbers(x, ...)

pcellnumbers(x, p)

pwcellnumbers(x, p)
}
\arguments{
\item{x}{Raster object}

\item{...}{arguments to extract}

\item{p}{Spatial polygons object}
}
\value{
tbl_df data frame
}
\description{
Title
}

