% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rastertable.r
\name{crop_}
\alias{crop_}
\alias{crop_.rastertable}
\title{Crop a rastertable with a 2D extent.}
\usage{
crop_(x, ...)

\method{crop_}{rastertable}(x, y, snap = "near", ...)
}
\arguments{
\item{x}{rastertable}

\item{...}{Additional arguments as for \code{\link{writeRaster}}}

\item{y}{Extent object, or any object from which an Extent object can be extracted (see Details)}

\item{snap}{Character. One of 'near', 'in', or 'out', for use with \code{\link{alignExtent}}}
}
\value{
rastertable
}
\description{
The cells that fall within \code{y} are determined and the table is filtered to these.
If no rows are left . . . something something.
}
\examples{
x <- voltab \%>\% crop_(c(0.5, 0.6, 0.1, 0.2))
}
\seealso{
\code{\link[raster]{crop}}
}
