% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestedtabula.r
\name{datatab}
\alias{datatab}
\title{library(raster)
library(maptools)
library(tibble)
library(dplyr)
data(wrld_simpl)
r <- rasterize(wrld_simpl, raster(nrow = 256, ncol = 512))
x <- rastertab(r)
plot(x, 6)}
\usage{
datatab(x)
}
\description{
library(raster)
library(maptools)
library(tibble)
library(dplyr)
data(wrld_simpl)
r <- rasterize(wrld_simpl, raster(nrow = 256, ncol = 512))
x <- rastertab(r)
plot(x, 6)
}

