% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer_extent.r
\name{as.double.Extent}
\alias{as.double.Extent}
\alias{as.integer.Extent}
\title{Atomic vector extent}
\usage{
\method{as.double}{Extent}(x, ...)

\method{as.integer}{Extent}(x, ...)
}
\arguments{
\item{x}{a \code{\link[raster]{extent}}}

\item{...}{unused}
}
\value{
numeric vector
}
\description{
Coerce a \code{\link[raster]{extent}} to an atomic vector of \code{c(xmin(x), xmax(x), ymin(x), ymax(x))}.
}
\details{
Note that \code{as.integer} results in truncation, see rasterOps for positive buffering.
}
\examples{
library(raster)
as.double(extent(0, 1, 0, 1))
as.numeric(extent(0, 1, 0, 1))
as.integer(extent(0, 1, 0, 1) + c(2.5, 27.877, 100, 999.1))
}
\seealso{
base::as.double
}
