% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rastertable.r
\name{extent_}
\alias{extent_}
\alias{extent_.rastertable}
\title{Title}
\usage{
extent_(x, ...)

\method{extent_}{rastertable}(x, ...)
}
\arguments{
\item{x}{rastertable}

\item{...}{ignored}
}
\value{
??
}
\description{
Return the extents of data values stored, as opposed to the full domain.
}
\examples{
extent_(voltab)
}
