% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extent_transforms.R
\name{extentFromCells}
\alias{extentFromCells}
\title{Extent from cells}
\usage{
extentFromCells(x, cells)
}
\arguments{
\item{x}{raster}

\item{cells}{cell numbers, relative to x}
}
\value{
raster extent
}
\description{
Use cell numbers to obtain their extent from a raster.
}
\details{
This is a missing funciton in the raster package for cell abstraction.
}
