% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extent_transforms.R
\name{index_extent}
\alias{index_extent}
\title{Index extent}
\usage{
index_extent(x, ex)
}
\arguments{
\item{x}{raster layer}

\item{ex}{extent}
}
\value{
extent object
}
\description{
Extent in index space.
}
\details{
Convert a geographic extent into purely index space.
}
\examples{
## the index extent is the rows/cols
index_extent(raster::raster(volcano), raster::extent(0, 1, 0, 1))

index_extent(raster::raster(volcano), raster::extent(0, 1, 0, .5))

}
