% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularaster-package.R
\name{oisst}
\alias{oisst}
\title{Optimally interpolated SST in near-native form.}
\format{
A data frame of sst values created from OISST data.
}
\description{
See data-raw/oisst.R in the source repository.  The file was
avhrr-only-v2.20170729.nc, its extent -180, 180, -90, 90 with
dimensions 1440x720 in the usual raster configuration.
}
\examples{
oisst
}
