% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyraster-package.r
\docType{data}
\name{polycano}
\alias{polycano}
\title{The raster volcano as polygons.}
\format{A \link\link{sp}{SpatialPolygonsDataFrame} with variables: \code{volcano_elevation}.}
\usage{
polycano
}
\description{
See data-raw/rastercano.r in the source repository.
}
\keyword{datasets}
