% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasa.r
\name{rasa.RasterLayer}
\alias{rasa.RasterLayer}
\title{Create a rasa.}
\usage{
\method{rasa}{RasterLayer}(x, ...)
}
\description{
The \code{rasa} method for \code{raster::RasterLayer}
}
\examples{
r <- raster::raster(volcano)
tr <- rasa(r)
## usual verbs
mutate(tr, a = 2)
## subsetting adds cell_index if not present
filter(tr, cell_value > 110)
slice(tr, c(1000, 1001))
plot(rasa(r) \%>\% filter(cell_value > 106) \%>\% slice(100:1000))

## more realish
library(maptools)
data(wrld_simpl)
r <- rasterize(wrld_simpl, raster(wrld_simpl, nrow = 360, ncol = 180))
tr <- rasa(r) \%>\% filter(!is.na(cell_value))
plot(tr \%>\% sample_n(15000))
plot(tr \%>\% filter(cell_index > 3e4))
}
