% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasa.r
\name{rasa}
\alias{rasa}
\title{Create a rasa.}
\usage{
rasa(x, ...)
}
\arguments{
\item{x}{input model}
}
\description{
A ... is a tibble storage mode for a "raster". This follows the model in the raster and sp packages where
raster cell values are stored in long form, in image-screen left-to-right, top-to-bottom order (same as used by GDAL).
}
\details{
We are similar to sp in that the columns are for variables, rows are for observations. In raster this is muddled, since
extra dimensions (e.g. time) are stored "wide" with a column for each "XY layer". Note that an RGB image is stored "wide"
in raster too, so ultimately it cannot have time-layer or z-layer varying RGB or other multi-column data (like U- V- velocity
vector components, or any multi-variable observations in higher than 2 dimensions).
}
