% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularaster-package.R
\name{raster-exports}
\alias{raster-exports}
\alias{extentFromCells}
\title{Re-exports from raster}
\description{
These functions are imported and then re-exported.
}
\details{
See \code{raster::\link[raster]{extentFromCells}} for details.
}
