% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularaster-package.r
\name{rastercano}
\alias{rastercano}
\title{The raster volcano.}
\format{A raster created from the \code{\link[datasets]{volcano}} data.}
\description{
See data-raw/rastercano.r in the source repository.
}
