% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rastertable.r
\name{rastertable}
\alias{rastertable}
\alias{rastertable.Raster}
\title{Create a rastertable.}
\usage{
rastertable(x, ...)

\method{rastertable}{Raster}(x, ...)
}
\arguments{
\item{x}{object as defined in `raster` package}

\item{...}{unused}
}
\value{
rastertable
}
\description{
A raster table is a long-form of data in the `raster` package.
}
\examples{
library(rastertable)
r <- raster(volcano)
rt <- rastertable(r)
}
