% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularaster-package.R
\name{sharkcano}
\alias{sharkcano}
\title{Sharkcano, the shark and the volcano.}
\format{
A data frame with 117843 rows and 2 variables:
\describe{
\item{\code{cell_}}{integer, cell index}
\item{\code{byte}}{integer, byte value of shark image pixels}
}
These are cell values on a grid that is 648x958.
}
\description{
This is just a free image off the internet.
The image was read in and all non-essential items dropped. The dimensions in \code{raster::raster} terms is stored in
\code{attr(sharkcano, "rasterdim")}.
}
\examples{
library(raster)
rd <- attr(sharkcano, "rasterdim")
rastershark <- raster(matrix(NA_integer_, rd[1], rd[2]))
rastershark[sharkcano$cell_] <- sharkcano$byte  ## byte, heh
## I present to you, Sharkcano!  (Just wait for the 3D version, Quadshark).
plot(rastercano)
contour(rastershark, add = TRUE, labels = FALSE)
plot(rastershark, col = "black")
## another way
plot(rastercano)
points(xyFromCell(rastershark, sharkcano$cell_), pch = ".")
}
\references{
This is the small version from here, see script in data-raw/sharkcano.r
http://www.freestockphotos.biz/stockphoto/16214
Thanks to @jennybc for pointers on finding free stuff: https://github.com/jennybc/free-photos
}
