% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-tabula.R
\name{stat_tabula}
\alias{stat_tabula}
\title{Leave data tabulastic!}
\usage{
stat_tabula(mapping = NULL, data = NULL, geom = "point",
  position = "identity", ..., show.legend = NA, inherit.aes = TRUE)
}
\description{
The tabula statistic is a kook.
}
\examples{
p <- ggplot(mtcars, aes(wt, mpg))
p + stat_tabula()

r <- raster::raster(volcano)
tr <- rasa(r)
#tr$x <- raster::xFromCell(r, seq_len(nrow(tr)))
#tr$y <- raster::yFromCell(r, seq_len(nrow(tr)))
library(ggplot2)
library(grid)
ggplot(tr, aes(x, y, fill = cell_value)) + geom_tabula()
library(tidyverse)
ggplot() +  geom_tabula(data = tr \%>\% filter(cell_value > 120, cell_value < 190), aes(x, y, fill = cell_value))

#library(elevatr)
#elev <- get_elev_raster(raster(extent(100, 150, -50, -20), crs = "+proj=longlat +datum=WGS84"), z = 3)
#saveRDS(rasa(elev), file = "rasa_data.rds", compress = "xz")
elev <- readRDS("rasa_data.rds")
ggplot() +  geom_tabula(data = elev \%>\% filter(cell_value < 120), aes(x, y, fill = cell_value))
}
