% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularaster-package.r
\docType{package}
\name{tabularaster}
\alias{tabularaster}
\alias{tabularaster-package}
\title{Tabular tools for raster, tidy tools for raster.}
\description{
Tabular tools for raster, tidy tools for raster.
}
