% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtopo.r
\name{topotab}
\alias{topotab}
\title{Return a table version of a raster.}
\usage{
topotab(ex, mapres = 1)
}
\arguments{
\item{ex}{Extent or vector c(xmin, xmax, ymin, ymax) see \code{\link[raster]{extent}}}

\item{mapres}{resolution in pixel size of desired map}
}
\value{
data frame with columns
}
\description{
Return a table for a built-in raster for a given extent at a given resolution.
}
\details{
The \code{mapres} defaults to 1, and cannot be lower than the native resolution, see \code{etopo2$temp}.
}
\examples{
topotab()
tab <- topotab(extent(120, 160, -70, -30), mapres = 0.5)
## avoid land?
tab <- dplyr::mutate(tab, topo = ifelse(topo < 0, topo, NA))
ggplot(tab)  + aes(x = x, y = y, fill = topo, width = width, height = height) +
scale_fill_gradient(low = "black", high = "white") + geom_raster()

g <- ggplot(tab)  + aes(x = x, y = y, fill = topo, width = width, height = height) + geom_raster()
g + scale_fill_gradient(low = "black", high = "white") 
g + scale_fill_gradientn(colours = topo.colors(100))

}

