% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rastertable.r
\docType{data}
\name{voltab}
\alias{voltab}
\title{Volcano raster map in rastertable form.}
\description{
This table is a raster with 61 columns and 87 rows, see \code{\link{volcano}}.
The values are stored in columns `cell_` and `value_` using the left-right top-bottom convention used by the raster package.
The raster skeleton of this is stored as an attribute `rtemplate`.
}
\details{
Created in /data-raw/.
}
\examples{
data(voltab)
plot(voltab, pch = 19, col = "#66666611")
}

