% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellnumbers.R
\name{cellnumbers}
\alias{cellnumbers}
\title{Extract cell numbers from a Raster object.}
\usage{
cellnumbers(x, query, ...)
}
\arguments{
\item{x}{Raster object}

\item{query}{Spatial object or matrix of coordinates}

\item{...}{arguments passed on}
}
\value{
tbl_df data frame
}
\description{
Currently only for single layer objects.
}

