% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellnumbers.R
\name{cellnumbers}
\alias{cellnumbers}
\title{Extract cell numbers from a Raster object.}
\usage{
cellnumbers(x, query, ...)
}
\arguments{
\item{x}{Raster object}

\item{query}{Spatial object or matrix of coordinates}

\item{...}{arguments passed on to \code{raster::cellFromPolygon} for polygon input}
}
\value{
tbl_df data frame
}
\description{
Currently only for single layer objects.
}
\details{
The dots argument can be used to set \code{weights=TRUE} for the polygon case, this is otherwise ignored.
}
\examples{
library(raster)
r <- raster(volcano)
cr <- cut(r,  pretty(values(r)))
p <- raster::rasterToPolygons(cr, dissolve = TRUE)
tt <- cellnumbers(cr, p)
library(dplyr)
tt \%>\% mutate(v = extract(r, cell_)) \%>\% 
group_by(object_) \%>\% 
summarize(mean(v)) 
head(pretty(values(r)), -1)
}
