% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularaster-package.R
\name{polycano}
\alias{polycano}
\title{The raster volcano as polygons.}
\format{
A \code{sp::SpatialPolygonsDataFrame} with variables: \code{volcano_elevation}.
}
\description{
See data-raw/rastercano.r in the source repository.
}
\examples{
exists("polycano")
}
