% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyraster-package.r
\docType{data}
\name{rastercano}
\alias{rastercano}
\title{The raster volcano.}
\format{A \link\link{raster}{raster} created from the \code{\link[datasets]{volcano}} data.}
\usage{
rastercano
}
\description{
See data-raw/rastercano.r in the source repository.
}
\keyword{datasets}
