% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularaster-package.R
\docType{package}
\name{tabularaster}
\alias{tabularaster}
\title{Tabular tools for raster}
\description{
Extract and index with raster tidy tools for raster.
}
\details{
Tabularaster includes these main functions.
\tabular{ll}{
\code{\link{as_tibble}} \tab convert raster data to data frame form, with control over output and form of dimension/coordinate columns \cr
\code{\link{cellnumbers}} \tab extract a data frame of query identifiers and cell,pixel index numbers \cr
\code{\link{decimate}} \tab fast and loose resizing of a raster to coarser resolution \cr
\code{\link{index_extent}} \tab build an extent in row column form, as opposed to coordinate value form \cr
}
}
